requirements-core.txt
# core dependencies (what the project really uses, without secondary dependencies) go here
dvc==3.5.1
hydra-core==1.3.2
imbalanced-learn==0.10.1
imblearn==0.0
missingno==0.5.2
optuna==3.2.0
pandas==2.0.2
shap==0.42.1
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
xgboost==1.7.5
openpyxl==3.1.2
fastparquet==2023.7.0
streamlit==1.24.1
